refact: simplify preference loading/saving
