+++ org.eclipse.tm4e.registry/src/main/java/org/eclipse/tm4e/registry/internal/GrammarRegistryManager.java
@@ -14,14 +14,11 @@ package org.eclipse.tm4e.registry.internal;
 import org.eclipse.core.runtime.IConfigurationElement;
 import org.eclipse.core.runtime.Platform;
 import org.eclipse.core.runtime.content.IContentType;
-import org.eclipse.core.runtime.preferences.InstanceScope;
 import org.eclipse.jdt.annotation.Nullable;
 import org.eclipse.tm4e.registry.GrammarDefinition;
 import org.eclipse.tm4e.registry.IGrammarDefinition;
 import org.eclipse.tm4e.registry.TMEclipseRegistryPlugin;
 import org.eclipse.tm4e.registry.XMLConstants;
-import org.eclipse.tm4e.registry.internal.preferences.PreferenceConstants;
-import org.eclipse.tm4e.registry.internal.preferences.PreferenceHelper;
 import org.osgi.service.prefs.BackingStoreException;
 
 /**
@@ -97,12 +94,8 @@ public final class GrammarRegistryManager extends AbstractGrammarRegistryManager
 	 * Load TextMate grammars from preferences.
 	 */
 	private void loadGrammarsFromPreferences() {
-		// Load grammar definitions from the
-		// "${workspace_loc}/metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.tm4e.registry.prefs"
-		final var prefs = InstanceScope.INSTANCE.getNode(TMEclipseRegistryPlugin.PLUGIN_ID);
-		final var json = prefs.get(PreferenceConstants.GRAMMARS, null);
-		if (json != null) {
-			final IGrammarDefinition[] definitions = PreferenceHelper.loadGrammars(json);
+		final var definitions = PreferenceHelper.loadGrammars();
+		if (definitions != null) {
 			for (final IGrammarDefinition definition : definitions) {
 				userCache.registerGrammarDefinition(definition);
 			}
@@ -111,11 +104,6 @@ public final class GrammarRegistryManager extends AbstractGrammarRegistryManager
 
 	@Override
 	public void save() throws BackingStoreException {
-		// Save grammar definitions in the
-		// "${workspace_loc}/metadata/.plugins/org.eclipse.core.runtime/.settings/org.eclipse.tm4e.registry.prefs"
-		final var json = PreferenceHelper.toJson(userCache.getDefinitions());
-		final var prefs = InstanceScope.INSTANCE.getNode(TMEclipseRegistryPlugin.PLUGIN_ID);
-		prefs.put(PreferenceConstants.GRAMMARS, json);
-		prefs.flush();
+		PreferenceHelper.saveGrammars(userCache.getDefinitions());
 	}
 }
