--- org.eclipse.tm4e.registry/src/main/java/org/eclipse/tm4e/registry/internal/preferences/PreferenceConstants.java
@@ -1,27 +1,0 @@
-/**
- * Copyright (c) 2015-2017 Angelo ZERR.
- * This program and the accompanying materials are made
- * available under the terms of the Eclipse Public License 2.0
- * which is available at https://www.eclipse.org/legal/epl-2.0/
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- * Angelo Zerr <angelo.zerr@gmail.com> - initial API and implementation
- */
-package org.eclipse.tm4e.registry.internal.preferences;
-
-/**
- * Preferences constants for grammar.
- */
-public final class PreferenceConstants {
-
-	public static final String GRAMMARS = "org.eclipse.tm4e.registry.grammars";
-
-	public static final String SCOPE_NAME_CONTENTTYPE_BINDINGS = "org.eclipse.tm4e.registry.scopeNameContentTypeBindings";
-
-	public static final String INJECTIONS = "org.eclipse.tm4e.registry.injections";
-
-	private PreferenceConstants() {
-	}
-}
