--- org.eclipse.tm4e.registry/src/main/java/org/eclipse/tm4e/registry/internal/preferences/package-info.java
@@ -1,4 +1,0 @@
-@NonNullByDefault
-package org.eclipse.tm4e.registry.internal.preferences;
-
-import org.eclipse.jdt.annotation.NonNullByDefault;
